+++ src/SuffixAutomaton.cpp
@@ -1,9 +1,9 @@
 #include "SuffixAutomaton.h"
 
 std::vector<State> st;
-int sz;
-int last;
-std::vector<std::vector<int>> inv_link;
+long long sz;
+long long last;
+std::vector<std::vector<long long>> inv_link;
 
 void sa_init(){
         st.clear();
@@ -13,7 +13,7 @@ void sa_init(){
 }
 
 void sa_extend(char c, long long current_pos){
-        int cur = sz++;
+        long long cur = sz++;
         st.emplace_back();
 
         st[cur].len = st[last].len + 1;
@@ -21,7 +21,7 @@ void sa_extend(char c, long long current_pos){
         st[cur].next.clear();
         st[cur].link = -1;
 
-        int p = last;
+        long long p = last;
         while(p != -1 && st[p].next.find(c) == st[p].next.end()){
                 st[p].next[c] = cur;
                 p = st[p].link;
@@ -29,11 +29,11 @@ void sa_extend(char c, long long current_pos){
         if(p == -1){
                 st[cur].link = 0;
         }else{
-                int q = st[p].next[c];
+                long long q = st[p].next[c];
                 if(st[p].len + 1 == st[q].len){
                         st[cur].link = q;
                 }else{
-                        int clone = sz++;
+                        long long clone = sz++;
                         st.emplace_back();
                         st[clone].len = st[p].len + 1;
                         st[clone].next = st[q].next;
@@ -50,16 +50,16 @@ void sa_extend(char c, long long current_pos){
 }
 
 void build_tree(){
-        inv_link.assign(sz, std::vector<int>());
-        for(int node = 1; node < sz; node++)
+        inv_link.assign(sz, std::vector<long long>());
+        for(long long node = 1; node < sz; node++)
                 inv_link[st[node].link].push_back(node);
 }
 
-void get_occurences(int v, std::vector<long long>& positions){
+void get_occurences(long long v, std::vector<long long>& positions){
         if(st[v].first_pos != -1){
                 positions.push_back(st[v].first_pos);
         }
-        for(const int&u : inv_link[v]){
+        for(const long long&u : inv_link[v]){
                 get_occurences(u, positions);
         }
 }
